Added: searching, user login, user logout, restricting pages

+++ phoenix/base/views.py
@@ -1,6 +1,11 @@
 from django.shortcuts import render, redirect
 from django.http import HttpResponse
-from .models import Room
+from django.db.models import Q
+from django.contrib.auth.decorators import login_required
+from django.contrib import messages
+from django.contrib.auth.models import User
+from django.contrib.auth import authenticate, login, logout
+from .models import Room, Topic
 from .forms import RoomForm
 
 
@@ -12,9 +17,50 @@ from .forms import RoomForm
 #     {'id': 5, 'name': '***** programistów HTMLa'},
 # ]
 
+def login_page(request):
+    if request.user.is_authenticated:
+        return redirect('home')
+
+
+    if request.method == "POST":
+        username = request.POST.get('username')
+        password = request.POST.get('password')
+
+        try:
+            user = User.objects.get(username=username)
+        except:
+            messages.error(request, "Wrong username or password")
+
+
+        user = authenticate(request, username = username, password = password)
+
+        if user is not None:
+            login(request, user)
+
+            return redirect('home')
+        else:
+            messages.error(request, "Wrong username or password")
+
+    context = {}
+    return render(request, 'base/login_register.html' , context )
+
+def logout_user(request):
+    logout(request)
+    return redirect('home')
+
+
 def home(request):
-    rooms = Room.objects.all()
-    context = {'rooms': rooms}
+    q = request.GET.get('q') if request.GET.get('q') is not None else ''
+    rooms = Room.objects.filter(
+        Q(topic__name__icontains=q) |
+        Q(name__icontains=q) |
+        Q(description__icontains=q)
+    )
+
+    topics = Topic.objects.all()
+    room_count = rooms.count()
+
+    context = {'rooms': rooms, 'topics': topics, "room_count" :room_count}
     return render(request, 'base/home.html', context)
 
 
@@ -23,7 +69,7 @@ def room(request, pk):
     context = {"room": room}
     return render(request, 'base/room.html', context)
 
-
+@login_required(login_url = "login")
 def createRoom(request):
     form = RoomForm()
     if request.method == "POST":
@@ -35,10 +81,13 @@ def createRoom(request):
     context = {'form': form}
     return render(request, 'base/room_form.html', context)
 
-
+@login_required(login_url = "login")
 def updateRoom(request, pk):
     room = Room.objects.get(id=pk)
     form = RoomForm(instance=room)
+
+    if request.user != room.host:
+        return HttpResponse("A co Ty tutaj chcesz?!")
 
     if request.method == "POST":
         form = RoomForm(request.POST, instance=room)
@@ -50,6 +99,8 @@ def updateRoom(request, pk):
 
 def deleteRoom(request, pk):
     room = Room.objects.get(id=pk)
+    if request.user != room.host:
+        return HttpResponse("A co Ty tutaj chcesz?!")
     if request.method == "POST":
         room.delete()
         return redirect('home')
